% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readline_but_better.R
\name{readline_choose}
\alias{readline_choose}
\title{User chooses from curated options}
\usage{
readline_choose(prompt = "please choose one of:", choices)
}
\arguments{
\item{prompt}{general prompt before the possible choices are listed (string)}

\item{choices}{vector describing the possible choices made available to user}
}
\value{
the value of the choice that the user selected
}
\description{
Asks end-user to make a selection between a curated set of possible choices. If user
}
\examples{
\dontrun{
readline_choose(
prompt = "Which fruit is your favourite?",
choices = c("Apples", "Oranges", "Pairs")
)
}
}
